+++ src/pathogena/__init__.py
@@ -1,3 +1,3 @@
 """The command line and Python client for EIT Pathogena."""
 
-__version__ = "2.2.1rc1"
+__version__ = "2.2.1"